\name{bcos}
\docType{data}
\alias{bcos}
\title{Breast Cosmesis Data}
\description{
 94 early breast cancer patients who had been treated with primary radiation therapy and adjuvant chemotherapy and with radiotherapy alone for the treatment. A patient due to monitored weekly or monthly for a clinically observable change in breast cancer status and may return with a changed status. For this case, we know the changed status is between ( \eqn{L}, \eqn{R} ].
}
\usage{
bcos
}
\format{
  \tabular{ll}{
    left:\tab breast cancer status return with a changed status after time \eqn{L}\cr
    right:\tab breast cancer status return with a changed status after time \eqn{R}\cr
    treatment:\tab Radiation = radiation therapy, RadChemo = radiation therapy with chemotherapy\cr
  }
}

\references{
Finkelstein DM and Wolfe RA. (1985).
A semiparametric model for regression analysis of interval-censored failure time data. \emph{Biometrics} 41, 933-945.
}
\keyword{datasets}
