\name{hdsd}
\docType{data}
\alias{hdsd}
\title{NASAs Hypobaric Decompression Sickness Data}
\description{
Time to onset of grade IV VGE was collected by volunteers who underwent denitrogenation test procedures before being uncovered to a hypobaric environment. This data set contains with 238 subjects and several covariates were recorded periodically monitor for a changed status is between ( \eqn{L}, \eqn{R} ].
}
\usage{
hdsd
}
\format{
  \tabular{ll}{
    Age:\tab recorded age (ranged from 20 to 54)\cr
    Sex:\tab (males = 1; females = 0)\cr
    TR360:\tab (ranged from 1.04 to 1.89)\cr
    Noadyn:\tab (ambulatory, Noadyn=1; lower body adynamic, Noadyn=0)\cr
    Left:\tab return with a grade IV VGE changed status after time \eqn{L}\cr
    Right:\tab return with a grade IV VGE changed status before time \eqn{R}\cr
  }
}

\references{
Conkin J, Bedahl SR and van Liew HD. (1992).
A computerized data bank of decompression sickness incidence in altitude chambers. \emph{Aviation Space Environ. Med} 63, 819-824.
}
\keyword{datasets}
