\name{print.thregI}
\alias{print.thregI}
\title{print method for thregI objects}
\description{
 Produces a printed summary of a fitted thregI function
 }
\usage{
\method{print}{thregI}(x, digits=max(options()$digits - 4, 3), ...)
}

\arguments{
  \item{x}{
    the result of a call to \code{thregI}
  }
  \item{digits}{significant digits to print}
  \item{\dots}{For future methods}
}

