\name{print.thregIcure}
\alias{print.thregIcure}
\title{print method for thregIcure objects}
\description{
 produces a printed summary of a fitted thregIcure function
 }
\usage{
\method{print}{thregIcure}(x, digits=max(options()$digits - 4, 3), ...)
}

\arguments{
  \item{x}{
    the result of a call to \code{thregIcure}
  }
  \item{digits}{significant digits to print}
  \item{\dots}{For future methods}
}

